refactor: ♻️ Move _check_min_nautobot_version_met to be done before integrations are added in Jobs dunder init.
